first commit
In this paper, we investigate the length of retirement in Brazil, which has a massive public pension program and an increasing life expectancy. We analyze the evolution of labor force participation rates and retirement from the late '70s and forecast future trends labor force participation and duration of retirement using the Lee-Carter model.